A simple backup script
Please, feel free to upgrade or modify the script if you want to!

This is a simple backup scripts that let's you save an entire directory to another location (not incremental yet!) and after that, ask you if you want to compress it on another location. Locations are variables, and they need to be modified.